requirements.txt
Django==5.0.3
environs==10.3.0
psycopg2-binary==2.9.9
channels[daphne]
python-json-logger
djangorestframework==3.15.1
aiohttp==3.9.3
requests==2.31.0
pyTelegramBotAPI
